requirements.txt
torch
torchvision
flask
pillow
pandas
matplotlib
pyngrok